requirements.txt
aioredis==2.0.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
bcrypt==4.0.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
click==8.1.6
cryptography==41.0.2
Deprecated==1.2.14
dnspython==1.16.0
ecdsa==0.18.0
exceptiongroup==1.1.2
fastapi==0.100.0
fastapi-limiter==0.1.5
h11==0.14.0
httpcore==0.17.3
idna==3.4
importlib-resources==6.0.0
jose==1.0.0
limits==3.5.0
packaging==23.1
passlib==1.7.4
pyasn1==0.5.0
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
pymongo==4.4.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
redis==4.6.0
rsa==4.9
six==1.16.0
slowapi==0.1.8
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.1
wrapt==1.15.0
